requirements.txt
beautifulsoup4==4.6.0
certifi==2017.4.17
chardet==3.0.4
idna==2.5
lxml==3.8.0
PyYAML==3.12
requests==2.18.1
urllib3==1.21.1